.emacs_d/lisp/bhj-set-key.el: do not mess with org-mode key binding bin/Darwin/after-check-out.sh: do not install my version of bash info if system is fixed already bin/adb-set-wifi-ap: fix wifi on the phone bin/delete-dup-files: delete dup files using md5sum bin/gerrit-push-review: do not challenge bin/gnus: let me know if the offlineimap is down bin/switch-to-vga: close hdmi

+++ bin/Darwin/after-check-out.sh
@@ -21,13 +21,13 @@ if test ! -d /usr/local/share/info; then
     sudo mkdir -p /usr/local/share/info
 fi
 
-sudo ln -s ~/doc/bash.info.gz /usr/local/share/info/ -f
-
-(
-    cd /usr/local/share/info/
-    sudo ginstall-info bash.info.gz /usr/local/share/info/dir
-)
-
+if test ! -e /usr/share/info/bash.info.gz; then
+    sudo ln -s ~/doc/bash.info.gz /usr/local/share/info/ -f
+    (
+        cd /usr/local/share/info/
+        sudo ginstall-info bash.info.gz /usr/local/share/info/dir
+    )
+fi
 #编译一些软件
 compile_beagrep
 compile_gtags
